Moved the Discriminator functionality out of the OpenApiSpecification class
