+++ core/src/main/kotlin/io/specmatic/conversions/OpenApiSpecification.kt
@@ -1152,93 +1152,6 @@ class OpenApiSpecification(
         }.toMap()
     }
 
-    data class Discriminator(private val discriminatorDetails: Map<String, Map<String, Pair<String, List<Schema<Any>>>>> = emptyMap()) {
-        fun isNotEmpty(): Boolean {
-            return discriminatorDetails.isNotEmpty()
-        }
-
-        val values: List<String>
-            get() {
-                return discriminatorDetails.entries.firstOrNull()?.value?.keys?.toList() ?: emptyList()
-            }
-
-        val key: String?
-            get() {
-                return discriminatorDetails.entries.firstOrNull()?.key
-            }
-
-        val schemas: List<Schema<Any>>
-            get() {
-                return discriminatorDetails.entries.flatMap {
-                    it.value.values.flatMap {
-                        it.second
-                    }
-                }
-            }
-
-        fun plus(newDiscriminatorDetails: Triple<String, Map<String, Pair<String, List<Schema<Any>>>>, Discriminator>?): Discriminator {
-            if(newDiscriminatorDetails == null)
-                return this
-
-            val (propertyName, valuesAndSchemas: Map<String, Pair<String, List<Schema<Any>>>>, newDiscriminator) = newDiscriminatorDetails
-
-            val updatedDiscriminatorDetails: Map<String, Map<String, Pair<String, List<Schema<Any>>>>> =
-                discriminatorDetails.plus(propertyName to valuesAndSchemas)
-
-            return this.copy(discriminatorDetails = updatedDiscriminatorDetails).plus(newDiscriminator)
-        }
-
-        fun plus(newDiscriminator: Discriminator): Discriminator {
-            return this.copy(discriminatorDetails = mergeMapOfMaps(discriminatorDetails, newDiscriminator.discriminatorDetails))
-        }
-
-        private fun <T> mergeMapOfMaps(
-            discriminatorDetails1: Map<String, Map<String, T>>,
-            discriminatorDetails2: Map<String, Map<String, T>>
-        ): Map<String, Map<String, T>> {
-            val keys = discriminatorDetails1.keys + discriminatorDetails2.keys
-
-            return keys.associateWith { key ->
-                val detail1 = discriminatorDetails1[key] ?: emptyMap()
-                val detail2 = discriminatorDetails2[key] ?: emptyMap()
-
-                (detail1 + detail2)
-            }
-        }
-
-        fun hasValueForKey(propertyName: String?): Boolean {
-            if(propertyName == null)
-                return false
-
-            return propertyName in discriminatorDetails
-        }
-
-        fun valueFor(propertyName: String): Pattern {
-            if(propertyName !in discriminatorDetails)
-                throw ContractException("$propertyName not found in discriminator details")
-
-            return discriminatorDetails.getValue(propertyName).firstNotNullOf { ExactValuePattern(StringValue(it.key), discriminator = true) }
-        }
-
-        fun explode(): List<Discriminator> {
-            return explode(discriminatorDetails)
-        }
-
-        private fun explode(discriminatorDetails: Map<String, Map<String, Pair<String, List<Schema<Any>>>>>): List<Discriminator> {
-            val propertyName = discriminatorDetails.keys.firstOrNull() ?: return listOf(Discriminator())
-
-            val discriminatorDetailsWithOneKeyLess = discriminatorDetails - propertyName
-
-            val valueOptionsWithSchemasForProperty = discriminatorDetails.getValue(propertyName)
-
-            return valueOptionsWithSchemasForProperty.flatMap { valueOption: Map.Entry<String, Pair<String, List<Schema<Any>>>> ->
-                explode(discriminatorDetailsWithOneKeyLess).map { discriminator ->
-                    discriminator.plus(Triple(propertyName, mapOf(valueOption.toPair()), Discriminator()))
-                }
-            }
-        }
-    }
-
     private fun isJsonInString(
         mediaType: MediaType, formFieldName: String?
     ) = if (mediaType.encoding.isNullOrEmpty()) false
